SDR: move to namespace nupic::sdr, allow copy-constructor
so SDR a = some_sdr; is now allowed.
This is an ugly commit, too many changes.

It mixes SDR moved under ns nupic::sdr; and some namespace fixes along
the way.

And need to replace &size, &dimensions with getter methods size(),
dimensions() to allow the copy constructor which is needed for likes of:
SDR a = sp.someFunctionReturningSDR();